+++ test/save_it/application_test.exs
@@ -1,8 +1,7 @@
 defmodule SaveIt.ApplicationTest do
   use ExUnit.Case, async: false
 
-  test "starts shared services without the Telegram bot in test" do
-    assert Process.whereis(SaveIt.TaskRunner.supervisor())
+  test "starts without the Telegram bot in test" do
     refute Process.whereis(SaveIt.BotSupervisor)
   end
 end